atulya_api/requirements.txt
# Requirements for atulya_api
django>=4.2.0
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
starlette>=0.27.0